Fixed RatePlaylist Functionality and Other Small Changes
RatePlaylist was not correctly preserving the logged in user's data. It
was overwriting their data with other users'. Fixed a similar issue in
ViewALlUsersPlaylists.

Within display stats, fixed a formatting issue that was causing single
digit seconds to not have a zero in front (e.g. instead of 3:01 it would
display 3:1)